extract favorites load/save methods to FileClient, separating with fetching methods which might change API Client in future
