+++ MyLibrary/Sources/DataClient/DataClient.swift
@@ -10,8 +10,6 @@ public struct DataClient {
   public var fetchWorkshop: @Sendable () throws -> Conference
   public var fetchSponsors: @Sendable () throws -> Sponsors
   public var fetchOrganizers: @Sendable () throws -> [Organizer]
-  public var loadFavorites: @Sendable () throws -> Favorites
-  public var saveFavorites: @Sendable (Favorites) throws -> Void
 }
 
 extension DataClient: DependencyKey {
@@ -41,16 +39,6 @@ extension DataClient: DependencyKey {
       let data = loadDataFromBundle(fileName: "organizers")
       let response = try jsonDecoder.decode([Organizer].self, from: data)
       return response
-    }, 
-    loadFavorites: {
-      guard let saveData = loadDataFromUserDefaults(key: "Favorites") else {
-        return .init(eachConferenceFavorites: [])
-      }
-      let response = try jsonDecoder.decode(Favorites.self, from: saveData)
-      return response
-    },
-    saveFavorites: { favorites in
-      saveDataToUserDefaults(favorites, as: "Favorites")
     }
   )
 
@@ -60,15 +48,6 @@ extension DataClient: DependencyKey {
     let data = try! Data(contentsOf: fileURL)
     return data
   }
-
-  static func saveDataToUserDefaults(_ favorites : Favorites, as key: String) {
-    let data = try? jsonEncoder.encode(favorites)
-    UserDefaults.standard.set(data, forKey: key)
-  }
-
-  static func loadDataFromUserDefaults(key: String) -> Data? {
-    return UserDefaults.standard.data(forKey: key)
-  }
 }
 
 let jsonDecoder = {
@@ -76,9 +55,3 @@ let jsonDecoder = {
   $0.keyDecodingStrategy = .convertFromSnakeCase
   return $0
 }(JSONDecoder())
-
-let jsonEncoder = {
-  $0.dateEncodingStrategy = .iso8601
-  $0.keyEncodingStrategy = .convertToSnakeCase
-  return $0
-}(JSONEncoder())
